pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "semantic-qa-gen"
version = "0.1.0"
description = "A Python library for generating high-quality question-answer pairs from PDF, DOCX, MD, and TXT files"
readme = "README.md"
authors = [
    {name = "Stephen Genusa", email = "github@genusa.com"},
]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: Implementation :: CPython",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: Linguistic",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
]
keywords = ["nlp", "question-generation", "ai", "llm", "question-answering", "semantic-analysis", "qa-pairs", "fine-tuning"]
dependencies = [
    "pyyaml>=6.0,<7.0",
    "httpx>=0.25.0,<0.28.0",
    "pydantic>=2.0.0,<3.0.0",
    "tqdm>=4.65.0,<5.0.0",
    "python-magic>=0.4.27,<0.5.0",
    "commonmark>=0.9.1,<0.10.0",
    "tiktoken>=0.4.0",
    "ruamel.yaml>=0.17.21,<0.19.0",
]
requires-python = ">=3.10"

[project.urls]
Documentation = "https://github.com/stephengenusa/semantic-qa-gen#readme"
Issues = "https://github.com/stephengenusa/semantic-qa-gen/issues"
Source = "https://github.com/stephengenusa/semantic-qa-gen"

[project.optional-dependencies]
pdf = [
    "pymupdf>=1.24.0,<1.25.0",
]
ocr = [
    "pytesseract>=0.3.10",
    "pillow>=9.5.0",
    "ftfy>=6.1.1",
]
advanced = [
    "scikit-learn>=1.2.0",
    "numpy>=1.24.0",
]
docx = [
    "python-docx>=1.0.0,<1.2.0",
]
formats = [
    "pymupdf>=1.24.0,<1.25.0",
    "python-docx>=1.0.0,<1.2.0",
    "pytesseract>=0.3.10",
    "pillow>=9.5.0",
    "ftfy>=6.1.1",
    "scikit-learn>=1.2.0",
    "numpy>=1.24.0",
]
openai = [
    "openai>=1.3.0,<2.0.0",
]
nlp = [
    "nltk>=3.8.1,<4.0.0",
]
full = [
    "pymupdf>=1.24.0,<1.25.0",
    "python-docx>=1.0.0,<1.2.0",
    "pytesseract>=0.3.10",
    "pillow>=9.5.0",
    "ftfy>=6.1.1",
    "scikit-learn>=1.2.0",
    "numpy>=1.24.0",
    "openai>=1.3.0,<2.0.0",
    "nltk>=3.8.1,<4.0.0",
    "rich>=13.3.5,<14.0.0",
]
dev = [
    "pytest>=7.3.1,<9.0.0",
    "pytest-asyncio>=0.21.0,<0.24.0",
    "pytest-cov>=4.1.0,<6.0.0",
    "black>=23.3.0,<25.0.0",
    "isort>=5.12.0,<6.0.0",
    "flake8>=6.0.0,<8.0.0",
    "mypy>=1.8.0,<2.0.0",
    "pre-commit>=3.0.0,<4.0.0",
    "mkdocs>=1.5.0,<1.7.0",
    "mkdocs-material>=9.1.15,<10.0.0",
    "types-pyyaml>=6.0.0,<7.0.0",
]

[project.scripts]
semantic-qa-gen = "semantic_qa_gen.cli.commands:main"

[tool.hatch.build.targets.wheel]
packages = ["semantic_qa_gen"]
